Move constants and delete config
