--- ui/src/config.js
@@ -1,4 +1,0 @@
-export const API_URL = process.env.REACT_APP_API_URL || '';
-export const DYNAMODB_ACCESS_KEY_ID = process.env.REACT_APP_DYNAMODB_ACCESS_KEY_ID;
-export const DYNAMODB_SECRET_ACCESS_KEY = process.env.REACT_APP_DYNAMODB_SECRET_ACCESS_KEY;
-export const DYNAMODB_TABLE = process.env.REACT_APP_DYNAMODB_TABLE || 'atlas-staging';
